backend-fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email-validator]==2.9.0
pydantic-settings==2.5.0
boto3==1.35.0
razorpay==1.4.2
